Redactored the decoder/encoder into codec
