+++ mmtf-codec/src/main/java/org/rcsb/mmtf/decoder/StructureDataToAdapter.java
