+++ mmtf-codec/src/main/java/org/rcsb/mmtf/decoder/package-info.java
