+++ mmtf-codec/src/main/java/org/rcsb/mmtf/encoder/EncoderUtils.java
