+++ mmtf-codec/src/main/java/org/rcsb/mmtf/encoder/WriterUtils.java
