+++ mmtf-codec/src/main/java/org/rcsb/mmtf/encoder/package-info.java
