+++ mmtf-codec/src/test/java/org/rcsb/mmtf/decoder/DummyTransferImpl.java
