+++ mmtf-codec/src/test/java/org/rcsb/mmtf/decoder/TestDefaultDecoder.java
