+++ mmtf-codec/src/test/java/org/rcsb/mmtf/encoder/DummyApiImpl.java
