+++ mmtf-codec/src/test/java/org/rcsb/mmtf/encoder/TestArrayEncoders.java
