+++ mmtf-codec/src/test/java/org/rcsb/mmtf/encoder/TestDefaultEncoder.java
